chore: extract life cycle to a separate file
Preparation for the future
